Renamed resources package to resource
