+++ src/main/java/com/fabbroniko/gamestatemanager/GameManager.java
@@ -8,8 +8,8 @@ import com.fabbroniko.environment.AudioManager;
 import com.fabbroniko.environment.Dimension;
 import com.fabbroniko.main.Drawable;
 import com.fabbroniko.main.IView;
-import com.fabbroniko.resources.ResourceManager;
-import com.fabbroniko.resources.domain.Level;
+import com.fabbroniko.resource.ResourceManager;
+import com.fabbroniko.resource.domain.Level;
 import com.fabbroniko.scene.AbstractScene;
 import com.fabbroniko.scene.GameScene;
 import com.fabbroniko.scene.LostScene;
