+++ src/main/java/com/fabbroniko/resource/ResourceManager.java
@@ -1,8 +1,8 @@
-package com.fabbroniko.resources;
+package com.fabbroniko.resource;
 
 import com.fabbroniko.error.ResourceNotFoundException;
-import com.fabbroniko.resources.domain.Background;
-import com.fabbroniko.resources.domain.Resource;
+import com.fabbroniko.resource.domain.Background;
+import com.fabbroniko.resource.domain.Resource;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import lombok.extern.log4j.Log4j2;
 
@@ -107,7 +107,7 @@ public class ResourceManager {
         log.trace("Loading audio clip from disk. Clip name {}", name);
 
         // Find the clip descriptor from the resource index
-        final Optional<com.fabbroniko.resources.domain.Clip> optResourceClip = resource.getClips()
+        final Optional<com.fabbroniko.resource.domain.Clip> optResourceClip = resource.getClips()
                 .stream()
                 .filter(c -> c.getName().equals(name))
                 .findAny();
