+++ src/main/java/com/fabbroniko/resource/domain/Map.java
@@ -1,4 +1,4 @@
-package com.fabbroniko.resources.domain;
+package com.fabbroniko.resource.domain;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlText;
 import lombok.Data;
