+++ src/main/java/com/fabbroniko/resource/domain/TileMap.java
@@ -1,4 +1,4 @@
-package com.fabbroniko.resources.domain;
+package com.fabbroniko.resource.domain;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlText;
