+++ src/main/java/com/fabbroniko/scene/GameScene.java
@@ -14,7 +14,7 @@ import com.fabbroniko.gameobjects.GameObjectBuilder;
 import com.fabbroniko.gameobjects.InvisibleBlock;
 import com.fabbroniko.gameobjects.Player;
 import com.fabbroniko.gamestatemanager.GameManager;
-import com.fabbroniko.resources.domain.Level;
+import com.fabbroniko.resource.domain.Level;
 
 import java.awt.*;
 import java.awt.event.KeyEvent;
